Refactor to isolate read stats functionality
