+++ bin-src/tasks/upload.test.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs-extra';
 import progressStream from 'progress-stream';
 
-import * as trimStatsFile from '../trim-stats-file';
+import * as readStatsFile from './read-stats-file';
 import { getDependentStoryFiles as dependentStoryFiles } from '../lib/getDependentStoryFiles';
 import { validateFiles, traceChangedFiles, uploadStorybook } from './upload';
 
@@ -91,7 +91,7 @@ describe('validateFiles', () => {
 });
 
 describe('traceChangedFiles', () => {
-  jest.spyOn(trimStatsFile, 'readStatsFile').mockReturnValueOnce(Promise.resolve(undefined));
+  jest.spyOn(readStatsFile, 'readStatsFile').mockReturnValueOnce(Promise.resolve(undefined));
 
   it('sets onlyStoryFiles on context', async () => {
     const deps = { 123: ['./example.stories.js'] };
