+++ bin-src/tasks/upload.ts
@@ -29,7 +29,7 @@ import {
 } from '../ui/tasks/upload';
 import { Context, Task } from '../types';
 import { exitCodes, setExitCode } from '../lib/setExitCode';
-import { readStatsFile } from '../trim-stats-file';
+import { readStatsFile } from './read-stats-file';
 
 const GetUploadUrlsMutation = `
   mutation GetUploadUrlsMutation($paths: [String!]!) {
