+++ bin-src/trace.ts
@@ -1,7 +1,7 @@
 import meow from 'meow';
 import { getDependentStoryFiles } from './lib/getDependentStoryFiles';
 import { Context } from './types';
-import { readStatsFile } from './trim-stats-file';
+import { readStatsFile } from './tasks/read-stats-file';
 
 /**
  * Utility to trace a set of changed file paths to dependent story files using a Webpack stats file.
